requirements.txt
django==2.0.5
djangorestframework==3.8.2
django-modeltranslation==0.13-beta1
django-ckeditor==5.4.0
django-admin-sortable2==0.6.19
django-taggit==0.22.2
django-taggit-serializer==0.1.5

six==1.11.0
Pillow==5.1.0